fix(BMEV-2789): change contract type

+++ src/main/java/com/stableapps/bookmapadapter/model/rest/InstrumentFutures.java
@@ -11,7 +11,7 @@ public class InstrumentFutures extends InstrumentGeneric {
     @JsonProperty("underlying_index")
     String underlyingIndex;
     @JsonProperty("contract_val")
-    int contractVal;
+    double contractVal;
     String listing;
     String delivery;
     @JsonProperty("trade_increment")
